Quittung Gider Kar Ciro Front Backend

+++ database/migrations/2022_10_26_164839_create_products_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('productName');
-            $table->integer('rentPrice');
-            $table->integer('buyPrice');
+            $table->float('rentPrice', 8, 2);
+            $table->float('buyPrice', 8, 2);
             $table->timestamps();
         });
     }
